requirements.txt
Django>=4.2
djangorestframework
dj-database-url
psycopg2-binary
gunicorn
whitenoise
Pillow
django-allauth
python-dotenv
